+++ NodeEditor/NodeEditorWindow/node_graphics_scene.py
